Add universal FSM for children no older than 7
From September 2014, all children in reception, year 1 and year 2 get
free school meals. This only applies in England and Scotland but this
commit imposes it for all parts of the UK.

No changes have been made in the case that free school meals are set
to be tapered away as part of the IS award (sys%extra%fsminappamt ==
.true.). This means universal free school meals are tapered away. This
may not be the best thing to do.