Correct error in benefit cap exemption
The benefit cap does not apply if the family is entitled to WTC. This
includes families who don't get any WTC but who would have got a WTC
award had their income not been too high. This is not what was
implemented: we required a positive WTC award. This has now been fixed.