Remove MatGrant from decision to take CCTaxRefund
TAXBEN does not set MatGrant to zero if CCTaxRefund is taken in
preference to UC/tax credits. The reason is: the woman was probably
not entitled to CCTaxRefund when the child was born because she was
not working. Therefore she could get MatGrant. We assume the one-off
MatGrant is spread across the 52 weeks of the year. Therefore the
family can still effectively be getting MatGrant even once the woman
has gone back to work and chosen to receive CCTaxRefund.